vectorstore and its int. test done; all int. tests done and refactored conftest
