+++ libs/astradb/tests/integration_tests/test_vectorstores.py
@@ -22,21 +22,19 @@ Required to run this test:
 import json
 import math
 import os
-import warnings
-from typing import Dict, Iterable, List, Optional, TypedDict
+from typing import Dict, Iterable, Optional
 
 import pytest
 from astrapy import Database
-from astrapy.db import AstraDB, AsyncAstraDB
+from astrapy.db import AstraDB
 from astrapy.info import CollectionVectorServiceOptions
 from langchain_core.documents import Document
-from langchain_core.embeddings import Embeddings
 
 from langchain_astradb.utils.astradb import SetupMode
 from langchain_astradb.vectorstores import AstraDBVectorStore
 
-from .conftest import _has_env_vars, AstraDBCredentials
-
+from ..conftest import ParserEmbeddings, SomeEmbeddings
+from .conftest import AstraDBCredentials, _has_env_vars
 
 # Faster testing (no actual collection deletions). Off by default (=full tests)
 SKIP_COLLECTION_DELETE = (
@@ -78,63 +76,6 @@ def is_nvidia_vector_service_available() -> bool:
             return False
     else:
         return False
-
-
-# Ad-hoc embedding classes:
-class SomeEmbeddings(Embeddings):
-    """
-    Turn a sentence into an embedding vector in some way.
-    Not important how. It is deterministic is all that counts.
-    """
-
-    def __init__(self, dimension: int) -> None:
-        self.dimension = dimension
-
-    def embed_documents(self, texts: List[str]) -> List[List[float]]:
-        return [self.embed_query(txt) for txt in texts]
-
-    async def aembed_documents(self, texts: List[str]) -> List[List[float]]:
-        return self.embed_documents(texts)
-
-    def embed_query(self, text: str) -> List[float]:
-        unnormed0 = [ord(c) for c in text[: self.dimension]]
-        unnormed = (unnormed0 + [1] + [0] * (self.dimension - 1 - len(unnormed0)))[
-            : self.dimension
-        ]
-        norm = sum(x * x for x in unnormed) ** 0.5
-        normed = [x / norm for x in unnormed]
-        return normed
-
-    async def aembed_query(self, text: str) -> List[float]:
-        return self.embed_query(text)
-
-
-class ParserEmbeddings(Embeddings):
-    """
-    Parse input texts: if they are json for a List[float], fine.
-    Otherwise, return all zeros and call it a day.
-    """
-
-    def __init__(self, dimension: int) -> None:
-        self.dimension = dimension
-
-    def embed_documents(self, texts: List[str]) -> List[List[float]]:
-        return [self.embed_query(txt) for txt in texts]
-
-    async def aembed_documents(self, texts: List[str]) -> List[List[float]]:
-        return self.embed_documents(texts)
-
-    def embed_query(self, text: str) -> List[float]:
-        try:
-            vals = json.loads(text)
-            assert len(vals) == self.dimension
-            return vals
-        except Exception:
-            print(f'[ParserEmbeddings] Returning a moot vector for "{text}"')
-            return [0.0] * self.dimension
-
-    async def aembed_query(self, text: str) -> List[float]:
-        return self.embed_query(text)
 
 
 @pytest.fixture(scope="function")
@@ -1234,7 +1175,10 @@ class TestAstraDBVectorStore:
         )
 
         # these invocations should just work without warnings
-        with pytest.warns(UserWarning) as rec_warnings:
+        # Note: mypy complains but the None argument seems the only way to
+        # count zero warnings without having to do global settings sucn as
+        # `warnings.simplefilter(...)`.
+        with pytest.warns(None) as rec_warnings:  # type: ignore[call-overload]
             AstraDBVectorStore(
                 collection_name="lc_default_idx",
                 embedding=embe,
@@ -1252,7 +1196,7 @@ class TestAstraDBVectorStore:
                 environment=astra_db_credentials["environment"],
                 metadata_indexing_exclude={"long_summary", "the_divine_comedy"},
             )
-        assert len(rec_warnings) == 1
+        assert len(rec_warnings) == 0
 
         # some are to throw an error:
         with pytest.raises(ValueError):
@@ -1349,7 +1293,10 @@ class TestAstraDBVectorStore:
         ).asimilarity_search("boo")
 
         # these invocations should just work without warnings
-        with pytest.warns(UserWarning) as rec_warnings:
+        # Note: mypy complains but the None argument seems the only way to
+        # count zero warnings without having to do global settings sucn as
+        # `warnings.simplefilter(...)`.
+        with pytest.warns(None) as rec_warnings:  # type: ignore[call-overload]
             def_store = AstraDBVectorStore(
                 collection_name="lc_default_idx",
                 embedding=embe,
@@ -1371,7 +1318,16 @@ class TestAstraDBVectorStore:
                 setup_mode=SetupMode.ASYNC,
             )
             await cus_store.aadd_texts(["All good."])
-        assert len(rec_warnings) == 0
+        """
+        the above leaves some unclosed network resources and fills the rec_warnings
+        with "unclosed <socket..." and "unclosed transport ...". Clean them out.
+        """
+        nonresource_warnings = [
+            r_w
+            for r_w in rec_warnings
+            if r_w.category != ResourceWarning
+        ]
+        assert len(nonresource_warnings) == 0
 
         # some are to throw an error:
         with pytest.raises(ValueError):
@@ -1436,12 +1392,12 @@ class TestAstraDBVectorStore:
                 environment=astra_db_credentials["environment"],
                 setup_mode=SetupMode.ASYNC,
             )
-        await leg_store.aadd_texts(["Triggering warning."])
+            await leg_store.aadd_texts(["Triggering warning."])
         assert len(rec_warnings) == 1
 
-        await database.drop_collection("lc_legacy_coll")
-        await database.drop_collection("lc_default_idx")
-        await database.drop_collection("lc_custom_idx")
+        await database.to_async().drop_collection("lc_legacy_coll")
+        await database.to_async().drop_collection("lc_default_idx")
+        await database.to_async().drop_collection("lc_custom_idx")
 
     @pytest.mark.skipif(
         os.environ.get("ASTRA_DB_ENVIRONMENT", "prod").upper() != "PROD",
